refactor(api): rename all eth to ethereum in api
